Add files via upload
This new version fixes a bug, namely the previous code did not check correctly for a draw. It checked for 42 turns, this is wrong since some of the turns may be wasted on trying to drop a token on a full column or on a non-existing column. So the current code actually checks if there are 42 bins full.